--- libs/accounts/CHANGELOG.md
@@ -1,88 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [0.2.13](https://github.com/thecointech/thecoin/compare/v0.2.12...v0.2.13) (2021-06-18)
-
-**Note:** Version bump only for package @thecointech/accounts
-
-
-
-
-
-## [0.2.12](https://github.com/thecointech/thecoin/compare/v0.2.11...v0.2.12) (2021-06-18)
-
-**Note:** Version bump only for package @thecointech/accounts
-
-
-
-
-
-## [0.2.11](https://github.com/thecointech/thecoin/compare/v0.2.10...v0.2.11) (2021-06-17)
-
-**Note:** Version bump only for package @thecointech/accounts
-
-
-
-
-
-## [0.2.10](https://github.com/thecointech/thecoin/compare/v0.2.9...v0.2.10) (2021-06-17)
-
-**Note:** Version bump only for package @thecointech/accounts
-
-
-
-
-
-## [0.2.9](https://github.com/thecointech/thecoin/compare/v0.2.8...v0.2.9) (2021-06-17)
-
-**Note:** Version bump only for package @thecointech/accounts
-
-
-
-
-
-## [0.2.8](https://github.com/thecointech/thecoin/compare/v0.2.7...v0.2.8) (2021-06-17)
-
-**Note:** Version bump only for package @thecointech/accounts
-
-
-
-
-
-## [0.2.7](https://github.com/thecointech/thecoin/compare/v0.2.6...v0.2.7) (2021-06-17)
-
-**Note:** Version bump only for package @thecointech/accounts
-
-
-
-
-
-## [0.2.6](https://github.com/thecointech/thecoin/compare/v0.2.1...v0.2.6) (2021-06-16)
-
-**Note:** Version bump only for package @thecointech/accounts
-
-
-
-
-
-## [0.2.2](https://github.com/thecointech/thecoin/compare/v0.1.29...v0.2.2) (2021-06-02)
-
-**Note:** Version bump only for package @thecointech/accounts
-
-
-
-
-
-## [0.2.1](https://github.com/thecointech/thecoin/compare/v0.2.0...v0.2.1) (2021-04-16)
-
-**Note:** Version bump only for package @thecointech/accounts
-
-
-
-
-
-# [0.2.0](https://github.com/thecointech/thecoin/compare/v0.1.29...v0.2.0) (2021-04-15)
-
-**Note:** Version bump only for package @thecointech/accounts
