+++ libs/nft-contract/tsconfig.build.json
@@ -7,6 +7,6 @@
   "exclude": ["**/*.test.*", "**/tests/**", "**/__mocks__/**", "**/*.stories.*"],
   "references": [
     { "path": "../logging/tsconfig.build.json" },
-    { "path": "../accounts/tsconfig.build.json" }
+    { "path": "../signers/tsconfig.build.json" }
   ]
 }
